server/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
psycopg2==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
email_validator==2.1.1
SQLAlchemy==2.0.28
scikit-learn
pandas==2.2.1
openai==1.16.2
requests==2.31.0
bs4==0.0.2
selenium==4.18.1
webdriver-manager==4.0.1
openai
tiktoken==0.6.0
html2text==2024.2.26